modal_sus_logic
