+++ app/src/components/modal/style.module.scss
@@ -1,0 +1,40 @@
+$bcInputs: #f5f5f5;
+$radiusButtonLogin: 10px;
+$bcButtonLogin: #8ca2d3;
+$sizeTextButtonAndLabels: 24px;
+$radiusButtonRegister: 10px;
+$bcButtonRegister: #8ca2d3;
+
+.modalWindow > div {
+	background-color: rgb(128, 128, 128);
+}
+
+.containerToModalButtons {
+	display: flex;
+	flex-direction: row;
+	justify-content: space-around;
+	align-items: center;
+}
+
+.submit_modal {
+	border: 1px solid black !important;
+	border-radius: $radiusButtonRegister;
+	width: 100px;
+	height: 45px;
+	font-size: $sizeTextButtonAndLabels;
+	border: none;
+}
+
+.cancel_modal {
+	background-color: $bcButtonRegister;
+	border-radius: $radiusButtonRegister;
+	width: 100px;
+	height: 45px;
+	font-size: $sizeTextButtonAndLabels;
+	border: none;
+}
+
+.modalText {
+	text-align: center;
+	font-size: 21px;
+}
